+++ src/test/kotlin/no/digdir/fdk/searchservice/utils/ApiTestContext.kt
@@ -30,7 +30,8 @@ abstract class ApiTestContext {
                 TEST_SEARCH_OBJECT_AND_HIT_ALL_FIELDS,
                 TEST_DATASET_HIT_ALL_FIELDS.toSearchObject(),
                 TEST_DATASET_HIT_IS_OPEN.toSearchObject(),
-                TEST_CONCEPT_HIT_ALL_FIELDS.toSearchObject()
+                TEST_CONCEPT_HIT_ALL_FIELDS.toSearchObject(),
+                TEST_DATASET_FILTERS.toSearchObject()
             ))
     }
